+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change history for ui-inventory
 
+## [12.0.15] (IN PROGRESS)
+
+* Ramsons CSP - Call number browse | Remove held by facet for ECS. Refs UIIN-3299.
+
 ## [12.0.14](https://github.com/folio-org/ui-inventory/tree/v12.0.14) (2025-03-26)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v12.0.13...v12.0.14)